chore: refactor encode service into transcribe service and vice versa. Removes cdk.out to lighten PRS
