--- cdk.out/cdk.out
@@ -1,1 +1,0 @@
-{"version":"11.0.0"}
